feat: start implement horisontal scaling
